Add files via upload
Added functionality to start button.
Updated inputs to hide when reset is clicked and reveal when start is clicked.
Scores count and reset.
Fixed uncaught RangeError. Caused by extra function call that was not needed.
Point limit to 3 per player implemented.
Winner is declared in webpage now.

TODO:
Add functionality to webpage to display the choices of both players.